requirements.txt
python-telegram-bot==22.2
python-dotenv
ytmusicapi
g4f
yt-dlp
aiohttp
aiofiles
curl-cffi
pillow
requests